Refactor: Rename google_client.py to client.py and changed display_name for the node
